+++ packages/elements/icon/glyph/objects/16/incident.js
@@ -1,0 +1,22 @@
+'use strict';
+
+Object.defineProperty(exports, "__esModule", {
+  value: true
+});
+
+var _extends = Object.assign || function (target) { for (var i = 1; i < arguments.length; i++) { var source = arguments[i]; for (var key in source) { if (Object.prototype.hasOwnProperty.call(source, key)) { target[key] = source[key]; } } } return target; };
+
+var _react = require('react');
+
+var _react2 = _interopRequireDefault(_react);
+
+var _index = require('../../../es5/index');
+
+var _index2 = _interopRequireDefault(_index);
+
+function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
+
+var Objects16IncidentIcon = function Objects16IncidentIcon(props) {
+  return _react2.default.createElement(_index2.default, _extends({ dangerouslySetGlyph: '<svg width="24" height="24" viewBox="0 0 24 24" focusable="false" role="presentation"><path d="M9.142 13h5.716l-.714-2H9.856l-.714 2zM8 15a1 1 0 0 0-1 1v1h10v-1a1 1 0 0 0-1-1H8zm2.57-6h2.86l-.488-1.368a1 1 0 0 0-1.884 0L10.57 9zM6 4h12a2 2 0 0 1 2 2v12a2 2 0 0 1-2 2H6a2 2 0 0 1-2-2V6a2 2 0 0 1 2-2z" fill="currentColor" fill-rule="evenodd"/></svg>' }, props));
+};
+exports.default = Objects16IncidentIcon;