+++ packages/elements/icon/src/components/Icon.js
@@ -103,7 +103,6 @@ export default class Icon extends Component<Props, {}> {
           primaryColor={primaryColor}
           secondaryColor={secondaryColor}
           size={size}
-          role="img"
           aria-label={this.props.label}
           dangerouslySetInnerHTML={{
             __html: Icon.insertDynamicGradientID(dangerouslySetGlyph),
@@ -118,7 +117,6 @@ export default class Icon extends Component<Props, {}> {
         primaryColor={primaryColor}
         secondaryColor={secondaryColor}
         size={size}
-        role="img"
         aria-label={this.props.label}
       >
         {Glyph ? <Glyph role="presentation" /> : null}