+++ packages/fabric/editor-confluence-transformer/src/parse.ts
@@ -349,6 +349,16 @@ function converter(
       case 'AC:TASK-LIST':
         return convertTaskList(schema, node) || unsupportedInline;
 
+      case 'AC:PLACEHOLDER':
+        const text = node.textContent;
+        if (text) {
+          return schema.nodes.placeholder.create({ text });
+        }
+        return null;
+
+      case 'FAB:ADF':
+        return convertADF(schema, node) || unsupportedInline;
+
       case 'PRE':
         return schema.nodes.codeBlock.create(
           { language: null },
@@ -411,6 +421,7 @@ function convertConfluenceMacro(
     ? parseDomNode(schema, richBodyNode).content
     : null;
   const plainTextBody = properties['ac:plain-text-body'] || '';
+  const schemaVersion = node.getAttributeNS(AC_XMLNS, 'schema-version');
 
   switch (macroName.toUpperCase()) {
     case 'CODE':
@@ -447,7 +458,6 @@ function convertConfluenceMacro(
       );
 
     case 'JIRA':
-      const schemaVersion = node.getAttributeNS(AC_XMLNS, 'schema-version');
       const { server, serverId, key: issueKey } = params;
 
       // if this is an issue list, render it as unsupported node
@@ -483,6 +493,7 @@ function convertConfluenceMacro(
           macroParams: getExtensionMacroParams(params),
           macroMetadata: {
             macroId: { value: macroId },
+            schemaVersion: { value: schemaVersion },
             placeholder: [
               {
                 data: { url: properties['fab:placeholder-url'] },
@@ -500,6 +511,7 @@ function convertConfluenceMacro(
           macroParams: getExtensionMacroParams(params),
           macroMetadata: {
             macroId: { value: macroId },
+            schemaVersion: { value: schemaVersion },
             placeholder: [
               {
                 data: { url: properties['fab:placeholder-url'] },
@@ -626,4 +638,16 @@ function convertTaskItem(schema: Schema, node: Element) {
   }
 
   return schema.nodes.taskItem.createChecked(attrs, nodes);
+}
+
+function convertADF(schema: Schema, node: Element) {
+  const str =
+    (node.textContent || '')[0] === '"'
+      ? node.textContent || ''
+      : `"${node.textContent}"`;
+  let json = JSON.parse(str);
+  if (typeof json === 'string') {
+    json = JSON.parse(json);
+  }
+  return schema.nodeFromJSON(json);
 }