+++ packages/fabric/editor-core/__tests__/ui/MentionPicker.tsx
@@ -8,7 +8,7 @@ describe('MentionPicker', () => {
   describe('Analytics', () => {
     let trackEvent;
     let component;
-    let componentInstance;
+    let componentInstance: MentionPicker;
 
     const mentionProvider = Promise.resolve(
       new MentionResource({
@@ -17,6 +17,10 @@ describe('MentionPicker', () => {
         productId: 'chat',
       }),
     );
+    const contextIdentifierProvider = Promise.resolve({
+      containerId: 'container-id',
+      objectId: 'object-id',
+    });
 
     const pluginKey = jest.fn() as any;
 
@@ -26,6 +30,7 @@ describe('MentionPicker', () => {
       component = shallow(
         <MentionPicker
           mentionProvider={mentionProvider}
+          contextIdentifierProvider={contextIdentifierProvider}
           pluginKey={pluginKey}
         />,
       );
@@ -42,6 +47,28 @@ describe('MentionPicker', () => {
       expect(trackEvent).toHaveBeenCalledWith(
         'atlassian.fabric.mention.picker.space',
         {},
+      );
+    });
+
+    it('should fire analytics in fireMentionInsert', () => {
+      (componentInstance as any).fireMentionInsertAnalytics({
+        id: 'AID-SECRET-@@@',
+        accessLevel: 'HIGH',
+      });
+      expect(trackEvent).toHaveBeenCalledWith(
+        'atlassian.fabric.mention.picker.insert',
+        {
+          containerId: 'container-id',
+          objectId: 'object-id',
+
+          mode: 'selected',
+          duration: 0,
+          accessLevel: 'HIGH',
+          isSpecial: false,
+          queryLength: 0,
+
+          mentionee: 'AID-SECRET-@@@',
+        },
       );
     });
   });