+++ packages/fabric/editor-core/examples/5-full-page.tsx
@@ -145,14 +145,21 @@ export default class Example extends React.Component<Props, State> {
               allowCodeBlocks={true}
               allowLists={true}
               allowTextColor={true}
-              allowTables={{ allowColumnResizing: true }}
+              allowTables={{
+                allowColumnResizing: true,
+                allowMergeCells: true,
+                allowNumberColumn: true,
+                allowBackgroundColor: true,
+                allowHeaderRow: true,
+                allowHeaderColumn: true,
+              }}
               allowJiraIssue={true}
               allowUnsupportedContent={true}
               allowPanel={true}
               allowExtension={true}
               allowRule={true}
               allowDate={true}
-              allowTemplatePlaceholders={true}
+              allowTemplatePlaceholders={{ allowInserting: true }}
               {...providers}
               media={{ provider: mediaProvider, allowMediaSingle: true }}
               placeholder="Write something..."