+++ packages/fabric/editor-core/src/editor/ui/Appearance/Comment.tsx
@@ -151,6 +151,7 @@ export default class Editor extends React.Component<
   private renderChrome = ({ maxContentSize }) => {
     const {
       editorView,
+      editorActions,
       eventDispatcher,
       providerFactory,
       contentComponents,
@@ -179,6 +180,7 @@ export default class Editor extends React.Component<
         <MainToolbar>
           <Toolbar
             editorView={editorView!}
+            editorActions={editorActions}
             eventDispatcher={eventDispatcher!}
             providerFactory={providerFactory!}
             appearance={this.appearance}
@@ -196,6 +198,7 @@ export default class Editor extends React.Component<
           {customContentComponents}
           <PluginSlot
             editorView={editorView}
+            editorActions={editorActions}
             eventDispatcher={eventDispatcher}
             providerFactory={providerFactory}
             appearance={this.appearance}