+++ packages/fabric/editor-core/src/ui/Task/task-item-with-providers.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { PureComponent, ReactElement } from 'react';
+import { Component, ReactElement } from 'react';
 import { ContextIdentifierProvider } from '@atlaskit/editor-common';
 import {
   ContentRef,
@@ -15,14 +15,14 @@ export interface Props {
   showPlaceholder?: boolean;
   children?: ReactElement<any>;
   taskDecisionProvider?: Promise<TaskDecisionProvider>;
-  contextIdentifierProvider?: Promise<any>;
+  contextIdentifierProvider?: Promise<ContextIdentifierProvider>;
 }
 
 export interface State {
-  resolvedContextProvider?: any;
+  resolvedContextProvider?: ContextIdentifierProvider;
 }
 
-export default class TaskItemWithProviders extends PureComponent<Props, State> {
+export default class TaskItemWithProviders extends Component<Props, State> {
   state: State = { resolvedContextProvider: undefined };
 
   componentWillMount() {
@@ -52,18 +52,13 @@ export default class TaskItemWithProviders extends PureComponent<Props, State> {
   }
 
   render() {
-    const {
-      taskDecisionProvider,
-      contextIdentifierProvider,
-      ...otherProps,
-    } = this.props;
+    const { contextIdentifierProvider, ...otherProps } = this.props;
     const { objectId, containerId } =
       this.state.resolvedContextProvider || ({} as ContextIdentifierProvider);
 
     return (
       <ResourcedTaskItem
         {...otherProps}
-        taskDecisionProvider={taskDecisionProvider}
         objectAri={objectId}
         containerAri={containerId}
       />