+++ packages/elements/analytics/CHANGELOG.md
@@ -1,5 +1,8 @@
 # @atlaskit/analytics
 
+## 2.4.4
+- [patch] updated the repository url to https://bitbucket.org/atlassian/atlaskit-mk-2 [1e57e5a](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/1e57e5a)
+
 ## 2.4.3
 - [patch] updated repository url to atlassian/atlaskit [9fb2698](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/9fb2698)