+++ packages/elements/single-select/CHANGELOG.md
@@ -1,5 +1,8 @@
 # @atlaskit/single-select
 
+## 3.2.4
+- [patch] updated the repository url to https://bitbucket.org/atlassian/atlaskit-mk-2 [1e57e5a](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/1e57e5a)
+
 ## 3.2.3
 - [patch] Packages Flow types for elements components [3111e74](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/3111e74)