+++ packages/fabric/renderer/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @atlaskit/renderer
 
+## 11.5.1
+- [patch] FS-1461 fixed rendererContext handling in TaskItem [6023540](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/6023540)
+
+## 11.5.0
+- [minor] FS-1461 objectAri and containerAri are optional in RendererContext [1b20296](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/1b20296)
+
+## 11.4.5
+- [patch] updated the repository url to https://bitbucket.org/atlassian/atlaskit-mk-2 [1e57e5a](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/1e57e5a)
+
+## 11.4.4
+
 ## 11.4.3